+++ src/tagger/phrase/index.js
@@ -1,15 +1,15 @@
 'use strict';
 const conditionPass = require('./00-conditionPass');
-// const verbPhrase = require('./01-verbPhrase');
+const verbPhrase = require('./01-verbPhrase');
 // const nounPhrase = require('./02-nounPhrase');
 // const AdjectivePhrase = require('./03-adjectivePhrase');
 //
-const phraseTag = function (Text) {
-  Text = conditionPass(Text);
-  // Text = verbPhrase(Text);
-  // Text = nounPhrase(Text);
-  // Text = AdjectivePhrase(Text);
-  return Text;
+const phraseTag = function (ts) {
+  ts = conditionPass(ts);
+  ts = verbPhrase(ts);
+  // ts = nounPhrase(ts);
+  // ts = AdjectivePhrase(ts);
+  return ts;
 };
 
 module.exports = phraseTag;